pyproject.toml
[tool.poetry]
name = "flask-tts-webapp"
version = "0.1.0"
description = ""
authors = ["Aadithya Sai <aadithyas0610@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask-magql = "^1.0.0"
gTTS = "^2.5.1"
flask-cors = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
